% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{classify}
\alias{classify}
\title{Classify data points}
\usage{
classify(data, samples)
}
\arguments{
\item{data}{a data frame with the data set used to estimate the \code{\link{hdpGLM}} model}

\item{samples}{the output of \code{\link{hdpGLM}}}
}
\description{
This function returns a data frame with the data points classified according to the estimation of cluster probabilities generated by the output of the function \code{\link{hdpGLM}}
}
