% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{dpGLM_get_occupied_clusters}
\alias{dpGLM_get_occupied_clusters}
\title{Get estimated \code{betas}}
\usage{
dpGLM_get_occupied_clusters(x)
}
\arguments{
\item{x}{the dpGLM or hdpGLM object returned by the function \link{hdpGLM}}
}
\description{
This function take the sample from the posterior and returns only the samples from the linear coefficients \code{beta} of the clusters with data points assigned to them.
}
