% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_hdpGLM_simulate.R
\name{dpGLM_simulateParameters}
\alias{dpGLM_simulateParameters}
\title{Simulated a Parameters used to create Data Sets from dpGLM model}
\usage{
dpGLM_simulateParameters(nCov, nCovj = NULL, K = NULL, pi = NULL,
  seed = NULL)
}
\arguments{
\item{nCov}{integer, the number of covariates of the GLM components}

\item{nCovj}{an integer indicating the number of covariates determining the average parameter of the base measure of the Dirichlet process prior}

\item{K}{integer, the number of clusters. If there are multiple contexts, K is the average number of clusters across contexts, and each context gets a number of clusters sampled from a Poisson distribution, except if \code{same.K} is \code{TRUE}.}

\item{pi}{either NULL or a vector with length K that add up to 1. If not NULL, it determines the mixture probabilities}

\item{seed}{a seed for \code{set.seed}}
}
\value{
The function returns a list with the parameters
        used to generate data sets from the dpGLM model.
        This list can be used in the function \code{hdpGLM_simulateData}
}
\description{
This function generates parameters that can be used to
simulated data sets from the Hierarchical
Dirichlet Process of Generalized Linear Model (dpGLM)
}
\examples{
parameters = dpGLM_simulateParameters(nCov=2, K=2) 

}
