% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{family.hdpGLM}
\alias{family.hdpGLM}
\title{Print GLM family}
\usage{
\method{family}{hdpGLM}(x, ...)
}
\arguments{
\item{x}{a \code{ndpGLM} object returned by the function
\code{hdpGLM}}

\item{...}{ignore}
}
\description{
Generic method to return the family of the GLM
}
