% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{hdpGLM_get_occupied_clusters}
\alias{hdpGLM_get_occupied_clusters}
\title{Get occupied Clusters}
\usage{
hdpGLM_get_occupied_clusters(x)
}
\arguments{
\item{x}{an \code{\link{dpGLM}} or \code{\link{hdpGLM}} object generated by the function \code{\link{hdpGLM}}}
}
\value{
The function returns the same object received, but it selects the occupoed clusters.
}
\description{
This function selects the clusters from the samples of the posterior distribution generated by the function \code{\link{hdpGLM}} that have data classified as belonging to them. See details.
}
\details{
The function \code{\link{hdpGLM}} generates samples from the posterior distribution of the hdpGLM model. That function also estimate the probability that a data point in the data set belongs to a specific cluster. If we classify the data points and assign them to the cluster they have the highest probability to belong to, it will give the clusters that are occupied with data points. This function selects such clusters from the posterior distribution and returns the sample from the posterior only with them.

\dontrun{
}
}
