% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{hdpGLM_match_clusters}
\alias{hdpGLM_match_clusters}
\title{Match Labels of Estimation and True Clusters}
\usage{
hdpGLM_match_clusters(samples, true)
}
\arguments{
\item{samples}{a \code{\link{dpGLM}} object, output of the function \code{\link{hdpGLM}}}

\item{true}{a list with the true values of the parameters that generate the data (See \code{\link{hdpGLM_simulateParameters}})}
}
\value{
The function returns a table with the labels of the clusters from the MCMC estimation provided by the function \code{\link{hdpGLM}} and the labels used to generate the data, which is in the column \code{Cluster.match} of the output. The output table also contains the true value of the linear coefficients that were used to generate the data.
}
\description{
This function matches the estimated clusters and the true values of the linear coefficients of the clusters.
}
