% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_dpGLM_simulate_data.R,
%   R/src_hdpGLM_simulate_data.R
\name{hdpGLM_simulateParameters}
\alias{hdpGLM_simulateParameters}
\alias{hdpGLM_simulateParameters}
\title{Simulated a Parameters used to create Data Sets from hdpGLM model}
\usage{
hdpGLM_simulateParameters(nCov, K, nCovj = 0, J = 1, pi = NULL,
  seed = NULL)

hdpGLM_simulateParameters(nCov, K, nCovj = 0, J = 1, pi = NULL,
  seed = NULL)
}
\arguments{
\item{nCov}{integer, the number of covariates of the
GLM components}

\item{K}{integer, the number of clusters}

\item{nCovj}{either NULL or an integer indicating the number of
covariates determining the average parameter of the
base measure of the Dirichlet process prior}

\item{J}{an integer representing the number of contexts}

\item{pi}{either NULL or a vector with length K that add up to 1.
If not NULL, it determines the mixture probabilities}

\item{seed}{a seed for \code{\link{set.seed}}}
}
\value{
The function returns a list with the parameters
        used to generate data sets from the hdpGLM model.
        This list can be used in the function \code{\link{hdpGLM_simulateData}}

The function returns a list with the parameters
        used to generate data sets from the hdpGLM model.
        This list can be used in the function \code{\link{hdpGLM_simulateData}}
}
\description{
This function generates parameters that can be used to
simulated data sets from the Hierarchical
Dirichlet Process of Generalized Linear Model (hdpGLM)

This function generates parameters that can be used to
simulated data sets from the Hierarchical
Dirichlet Process of Generalized Linear Model (hdpGLM)
}
\examples{
parameters = hdpGLM_simulateParameters(nCov=2) 

\dontrun{
}
parameters = hdpGLM_simulateParameters(nCov=2) 

\dontrun{
}
}
