% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{mcmc_info.hdpGLM}
\alias{mcmc_info.hdpGLM}
\title{mcmc}
\usage{
mcmc_info.hdpGLM(x, ...)
}
\arguments{
\item{x}{a \code{hdpGLM} object returned by the function
\code{hdpGLM}}

\item{...}{ignore}
}
\description{
Generic method to return the MCMC information
}
