% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{nclusters}
\alias{nclusters}
\title{nclusters}
\usage{
nclusters(object)
}
\arguments{
\item{object}{a \code{dpGLM} object returned by the function \code{hdpGLM}}
}
\description{
This function returns the number of clusters found in the estimation
}
