% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{nobs.dpGLM}
\alias{nobs.dpGLM}
\title{nobs}
\usage{
\method{nobs}{dpGLM}(x, ...)
}
\arguments{
\item{x}{a \code{dpGLM} object returned by the function
\code{hdpGLM}}

\item{...}{ignore}
}
\description{
Generic method to return the size of the data
}
