% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{plot.dpGLM}
\alias{plot.dpGLM}
\title{Defaul plot for class dpGLM}
\usage{
\method{plot}{dpGLM}(x, K = 1, separate = F, hpd = T,
  true_parameters = NULL, plot.hist = T, plot.cluster.freq = T,
  layout = NULL, title = NULL, prop.time.active = 0.95, adjust = 0.8,
  ...)
}
\arguments{
\item{x}{a dpGLM object with the samples from generated by \code{\link{hdpGLM}}}

\item{K}{integer, the number of estimated clusters to plot}

\item{separate}{boolean, if \code{TRUE} the coefficients of the \code{K} first clusters will be displayed in a separate panel in the plot. If \code{FALSE}, only the beta coefficient will be displayed in a separate panel in the plot}

\item{hpd}{boolean, if \code{TRUE} the and \code{separate=T}, the .95 HPD lines will be displayed.}

\item{true_parameters}{either \code{NULL} or a list with the true parameter values if they are known. The list must be as prescribed for the parameter \code{parameters} of the function \code{\link{hdpGLM_simulateData}}}

\item{plot.hist}{boolean, if \code{TRUE} the histogram is shown together with the densities of the parameters}

\item{plot.cluster.freq}{boolean, if \code{TRUE} it generates a bar plot shows the clusters and how frequent data points were classified in that cluster during the estimation of the model}

\item{layout}{either \code{NULL} or a two-dimensional vector with integers indicating the number of rows and coluns to plot. If \code{NULL}, the grid is provided}

\item{title}{string, the title of the plot}

\item{prop.time.active}{a number between 0 and 1. When \code{separate=F} and \code{plot.cluster.freq=T}, it will count the number of cluster that was active in the provided proportion of the time}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}.
    This makes it easy to specify values like \sQuote{half the default}
    bandwidth.}

\item{...}{arguments in \code{tag = value} form, or a list of tagged
    values.  The tags must come from the names of graphical parameters
    described in the \sQuote{Graphical Parameters} section.}
}
\value{

}
\description{
This function generate desity plots with the posterior distribution generated by the function \code{\link{hdpGLM}}
}
\examples{
dt    = hdpGLM_simulateData(n=5000,nCov=4, K=5, family='gaussian')
mcmc    = list(burn.in = 0,  n.iter = 2000)
samples = hdpGLM(y~., data=dt$data, mcmc=mcmc,
                 family='gaussian', n.display=30, K=100)

plot(samples, K=5, true_parameters=data$parameters)

\dontrun{
}
}
