% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{plot.dpGLM_data}
\alias{plot.dpGLM_data}
\title{Plot simulated data from hdpGLM model}
\usage{
\method{plot}{dpGLM_data}(sim_data, title = NULL, ...)
}
\arguments{
\item{title}{string, the title of the plot}

\item{x}{a dpGLM_data object, generated by the funciton  \code{\link{hdpGLM_simulateData}}}
}
\description{
This is a generic function applied to the class \code{dpGLM_data}, which is a list with a data set generated by the function \code{\link{hdpGLM_simulateData}}. It creates a plot describing the data set generated by that function.
}
\examples{

\dontrun{
}
}
