% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{plot_beta_sim}
\alias{plot_beta_sim}
\title{Plot simulated data}
\usage{
plot_beta_sim(data, w.idx, ncol = NULL)
}
\arguments{
\item{data}{the output of the function \code{\link{hdpGLM_simulateData}}}

\item{w.idx}{integer, the index of the context level covariate the plot}

\item{ncol}{integer, the number of columns in the grid of the plot}
}
\description{
Create a plot with the beta sampled from its distribution, as a function of context-level feature $W$
}
