% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{predict.hdpGLM}
\alias{predict.hdpGLM}
\title{hdpGLM Predicted values}
\usage{
\method{predict}{hdpGLM}(object, new_data = NULL, ...)
}
\arguments{
\item{object}{outcome of the function hdpLGM}

\item{new_data}{data frame with the values of the covariates that are going to be used to generate the predicted/fitted values. The posterior mean is used to create the predicted values}

\item{...}{\code{family} : a string with the family of the output variable: \code{gaussian} (default), \code{binomial}, etc...}
}
\value{
It returns a data.frame with the fitted values for the outcome variable, which are produced using the estimated posterior expectation of the linear coefficients \code{beta}.
}
\description{
Function returns the predicted (fitted) values of the outcome variable using the estimated posterior expectation of the linear covariate betas produced by the \code{hdpGLM} function
}
