% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{print.hdpGLM_data}
\alias{print.hdpGLM_data}
\title{Print}
\usage{
\method{print}{hdpGLM_data}(x, ...)
}
\arguments{
\item{x}{a \code{hdpGLM_data} object returned by the function
\code{hdpGLM_simulateData}}

\item{...}{ignore}
}
\value{
returns a summary of the simulated data
}
\description{
Generic method to print the output of the \code{hdpGLM_simulateData} function
}
