% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{summary.dpGLM}
\alias{summary.dpGLM}
\title{Default Summary function for dpGLM class}
\usage{
\method{summary}{dpGLM}(x, nK = NULL, HPD.prob = 0.95, burn.in = NULL,
  only.occupied.clusters = TRUE)
}
\arguments{
\item{x}{a MCMC object \code{\link{coda}} returned by the \code{\link{dpGLM}} function}

\item{nK}{either \code{NULL} or an integer representing the number of clusters to display the summary statistics. The clusters displayed are ordered by the proportion of time they were active during the MCMC simulation. If \code{NULL}, the function displays only the clusters that contains data points (see Details)}

\item{HPD.prob}{a number between 0 and 1 (default is .95). It is used to display the HPD of the posterior distribution}

\item{only.occupied.clusters}{boolean, if \code{TRUE} it shows the only the densities of the clusters that actually have data points assigned to it with high probability}
}
\description{
This function provides a summary of the MCMC samples from the dpGLM model
}
\details{
Data points are assigned to clusters according to the highest estimated probability of belonging to that cluster
}
