% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{summary.dpGLM_data}
\alias{summary.dpGLM_data}
\title{Summary dpGLM data}
\usage{
\method{summary}{dpGLM_data}(x, ...)
}
\arguments{
\item{x}{an object of the class dpGLM_data}

\item{...}{ignored}
}
\value{
The function returns a list with the summary of the data produced by the standard summary function and a \code{data.frame} with the true values of beta for each cluster.
}
\description{
This function summarizes the data and parameters used to generate the data using the function hdpLGM.
}
