% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{summary.hdpGLM_data}
\alias{summary.hdpGLM_data}
\title{Summary}
\usage{
\method{summary}{hdpGLM_data}(object, ...)
}
\arguments{
\item{object}{an object of the class \code{hdpGLM_data}, which is produced by the function \code{hdpGLM_simulate_data}}

\item{...}{ignored}
}
\value{
It returns a list with three elements. The first is a summary of the data, the second a data_frame with the linear coefficients \code{beta} and their values used to generate the data, and the third element is also a data_frame with the true values of \code{tau} used to generate the \code{betas}.
}
\description{
This functions summarizes the data simulated by the function \code{hdpGLM_simulate_data}
}
