% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpGLM_data.R
\docType{data}
\name{welfare}
\alias{welfare}
\title{Fake data set with 2000 observations}
\format{
A data frame with 2000 rows and 4 variables:
\describe{
  \item{support}{support for welfare policies}
  \item{inequality}{levels of inequality in the neighborhood}
  \item{income}{individual-level income}
  \item{ideology}{individual-level ideology}
}
}
\source{
Simulated data
}
\usage{
welfare
}
\description{
A dataset containing simulated data about public opinion
}
\keyword{datasets}
