% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpGLM_data.R
\docType{data}
\name{welfare2}
\alias{welfare2}
\title{Fake data set with 2000 observations}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
  \item{support}{support for welfare policies}
  \item{inequality}{levels of inequality in the neighborhood}
  \item{income}{individual-level income}
  \item{ideology}{individual-level ideology}
  \item{country}{country label or index}
  \item{gap}{country-level gender gap in country's provision of public good}
}
}
\source{
Simulated data
}
\usage{
welfare2
}
\description{
A dataset containing simulated data about public opinion in different
countries
}
\keyword{datasets}
