% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{coef.dpGLM}
\alias{coef.dpGLM}
\title{Default function to extract coefficients from dpGLM fitted model}
\usage{
\method{coef}{dpGLM}(object, ...)
}
\arguments{
\item{object}{a \code{dpGLM} object returned by the function \code{hdpGLM}}

\item{...}{The additional parameters accepted are:}
}
\description{
This function gives the posterior mean of the coefficients
}
