% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{hdpGLM_classify}
\alias{hdpGLM_classify}
\title{Deprecated}
\usage{
hdpGLM_classify(data, samples)
}
\arguments{
\item{data}{a data frame with the data set used to estimate the \code{\link{hdpGLM}} model}

\item{samples}{the output of \code{\link{hdpGLM}}}
}
\description{
Deprecated
}
