% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_dpGLM_simulate_data.R
\name{hdpGLM_simulateData}
\alias{hdpGLM_simulateData}
\title{Simulated a Data Set from hdpGLM model}
\usage{
hdpGLM_simulateData(n, K, nCov, nCovj = NULL, parameters = NULL,
  pi = NULL, family, seed = sample(1:777, 1))
}
\arguments{
\item{n}{integer, the sample size of the data}

\item{K}{integer, the number of clusters}

\item{nCov}{integer, the number of covariates of the
GLM components}

\item{nCovj}{either NULL or an integer indicating the number of
covariates determining the average parameter of the
base measure of the Dirichlet process prior}

\item{parameters}{either NULL or a list with the parameters
to generate the model. If not NULL, it must
contain a sublist name beta, a vector named tau,
and a vector named pi. The sublist beta must be
a list of vectors, each one with size nCov+1
to be the coefficients of the GLM mixtures
components that will generate the data.
The vector tau must be present only if nCovj
is not NULL, in which case it must be a vector
of size nCov+1 with the linear coefficients of
the average of the base measure. The vector pi
must add up to 1 and have length K.}

\item{pi}{either NULL or a vector with length K that add up to 1.
If not NULL, it determines the mixture probabilities}

\item{family}{a character with either 'gaussian', 'binomial', or 'multinomial'.
It indicates the family of the GLM components of the mixture model.}

\item{seed}{a seed for \code{\link{set.seed}}}
}
\value{
The function returns a list with the data set, the parameters,
        the vector Z indicating the cluster of each observation, and
        the number of clusters.
}
\description{
This function generates a data set from the Hierarchical
Dirichlet Process of Generalized Linear Model (hdpGLM)
}
\examples{
data = hdpGLM_simulateData(n=2000, K=2, nCov=0, family='gaussian')
\dontrun{
}
}
