% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_dpGLM_simulate_data.R
\name{hdpGLM_simulateParameters}
\alias{hdpGLM_simulateParameters}
\title{Simulated a Parameters used to create Data Sets from hdpGLM model}
\usage{
hdpGLM_simulateParameters(nCov, nCovj = NULL, K, pi = NULL, seed = NULL)
}
\arguments{
\item{nCov}{integer, the number of covariates of the
GLM components}

\item{nCovj}{either NULL or an integer indicating the number of
covariates determining the average parameter of the
base measure of the Dirichlet process prior}

\item{K}{integer, the number of clusters}

\item{pi}{either NULL or a vector with length K that add up to 1.
If not NULL, it determines the mixture probabilities}

\item{seed}{a seed for \code{\link{set.seed}}}
}
\value{
The function returns a list with the parameters
        used to generate data sets from the hdpGLM model.
        This list can be used in the function \code{\link{hdpGLM_simulateData}}
}
\description{
This function generates parameters that can be used to
simulated data sets from the Hierarchical
Dirichlet Process of Generalized Linear Model (hdpGLM)
}
\examples{
parameters = hdpGLM_simulateParameters(nCov=2) 

\dontrun{
}
}
