% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_hdpGLM_simulate_data.R
\name{hdpGLM_simulateParameters}
\alias{hdpGLM_simulateParameters}
\title{Simulated a Parameters used to create Data Sets from hdpGLM model}
\usage{
hdpGLM_simulateParameters(nCov, K = NULL, nCovj = NULL, J = NULL,
  pi = NULL, same.K, seed = NULL, context.effect = NULL,
  same.clusters.across.contexts, context.dependent.cluster)
}
\arguments{
\item{nCov}{integer, the number of covariates of the
GLM components}

\item{K}{integer, the number of clusters. If there are multiple contexts, K is the average number of clusters across contexts, and each context gets a number of clusters sampled from a Poisson distribution, except if \code{same.K} is \code{TRUE}.}

\item{nCovj}{an integer indicating the number of
covariates determining the average parameter of the
base measure of the Dirichlet process prior}

\item{J}{an integer representing the number of contexts}

\item{pi}{either NULL or a vector with length K that add up to 1.
If not NULL, it determines the mixture probabilities}

\item{same.K}{boolean, used when data is sampled from more than one context. If \code{TRUE} all contexts get the same number of clusters. If \code{FALSE}, each context gets a number of clusters sampled from a Poisson distribution with expectation equals to \code{K} (current not implemented)}

\item{seed}{a seed for \code{\link{set.seed}}}

\item{context.effect}{either \code{NULL} or a two dimensional integer vector. If it is \code{NULL}, all the coefficients (\code{beta}) of the individual level covariates are functions of context-level features (\code{tau}). If it is not \code{NULL}, the first component of the vector indicates the index of the lower level covariate (\code{X}) whose linear effect \code{beta} depends on context (\code{tau}) (0 is the intercept). The second component indicates the index context-level covariate (\code{W}) whose linear coefficient (\code{tau}) is non-zero.}

\item{same.clusters.across.contexts}{boolean, if \code{TRUE} all the contexts will have the same number of clusters AND each cluster will have the same coefficient \code{beta}.}

\item{context.dependent.cluster}{integer, indicates which cluster will be context-dependent. If \code{zero}, all clusters will be context-dependent}
}
\value{
The function returns a list with the parameters
        used to generate data sets from the hdpGLM model.
        This list can be used in the function \code{\link{hdpGLM_simulateData}}
}
\description{
This function generates parameters that can be used to
simulated data sets from the Hierarchical
Dirichlet Process of Generalized Linear Model (hdpGLM)
}
\examples{
parameters = hdpGLM_simulateParameters(nCov=2, K=2, nCovj=3, J=20,
                same.clusters.across.contexts=FALSE, context.dependent.cluster=0) 

}
