% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{plot.dpGLM}
\alias{plot.dpGLM}
\title{Defaul plot for class dpGLM}
\usage{
\method{plot}{dpGLM}(x, K = NULL, separate = FALSE, hpd = TRUE,
  true_parameters = NULL, plot.hist = TRUE,
  plot.cluster.classification = TRUE, layout = NULL, title = NULL,
  prop.time.active = 0.95, adjust = NULL, only.occupied.clusters = TRUE,
  include.legend = TRUE, yaxis = T, ...)
}
\arguments{
\item{x}{a dpGLM object with the samples from generated by \code{\link{hdpGLM}}}

\item{K}{either \code{NULL} or an integer indicating the number of estimated clusters to plot}

\item{separate}{boolean, if \code{TRUE} the coefficients of the \code{K} first clusters will be displayed in a separate panel in the plot. If \code{FALSE}, only the beta coefficient will be displayed in a separate panel in the plot}

\item{hpd}{boolean, if \code{TRUE} the and \code{separate=T}, the .95 HPD lines will be displayed.}

\item{true_parameters}{either \code{NULL} or a list with the true parameter values if they are known. The list must be as described for the parameter \code{parameters} of the function \code{\link{hdpGLM_simulateData}}}

\item{plot.hist}{boolean, if \code{TRUE} the histogram is shown together with the densities of the parameters}

\item{layout}{either \code{NULL} or a two-dimensional vector with integers indicating the number of rows and coluns to plot. If \code{NULL}, the grid is provided}

\item{title}{string, the title of the plot}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}.
    This makes it easy to specify values like \sQuote{half the default}
    bandwidth.}

\item{only.occupied.clusters}{boolean, if \code{TRUE} it shows the only the densities of the clusters that actually have data points assigned to it with high probability}

\item{include.legend}{boolean, if \code{TRUE} a legend is displayed in the plot identifying the clusters. Used only if \code{separate=FALSE}.}

\item{yaxis}{boolean, indicates if show y axis in the plot or not 
# @param prop.time.active a number between 0 and 1. When \code{separate=F} and \code{plot.cluster.freq=T}, it will count the number of cluster that was active in the provided proportion of the time
# @param select_perc_time_active a number between 0 and 100 indicating the cluster to be displayed according to the percentage of the time it was active during the MCMC estimation
# @param omit.irrelevant.clusters boolean, if \code{TRUE} it will omit the irrelevant clusters, that is, those cluster whose vector of linear coefficients are not indistinguishable from the zero vector (default is \code{TRUE})}

\item{...}{arguments in \code{tag = value} form, or a list of tagged
    values.  The tags must come from the names of graphical parameters
    described in the \sQuote{Graphical Parameters} section.}

\item{plot.cluster.classfication}{boolean, if \code{TRUE} it generates a bar plot that shows the clusters and the percentage of the data points classified in that cluster}
}
\value{

}
\description{
This function generate desity plots with the posterior distribution generated by the function \code{\link{hdpGLM}}
}
\examples{
dt    = hdpGLM_simulateData(n=5000,nCov=4, K=5, family='gaussian')
mcmc    = list(burn.in = 0,  n.iter = 2000)
samples = hdpGLM(y~., data=dt$data, mcmc=mcmc,
                 family='gaussian', n.display=30, K=100)

plot(samples, true_parameters=data$parameters)

\dontrun{
}
}
