% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{print.dpGLM}
\alias{print.dpGLM}
\title{Print}
\usage{
\method{print}{dpGLM}(x, ...)
}
\arguments{
\item{x}{a \code{dpGLM} object returned by the function \code{hdpGLM}}

\item{...}{ignore}
}
\value{
returns a summary of the posterior distribution of the parameters
}
\description{
Generic method to print the output of a \code{hdpGLM} of \code{dpGLM} objects
}
