% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{print.hdpGLM}
\alias{print.hdpGLM}
\title{Print}
\usage{
\method{print}{hdpGLM}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{hdpGLM} generted by the function \link{hdpGLM}}

\item{...}{ignore}
}
\value{
returns a summary of the posterior distribution of the parameters
}
\description{
Generic method to print the output of a \code{hdpGLM} of \code{dpGLM} objects
}
