% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_summaries.R
\name{summary.dpGLM}
\alias{summary.dpGLM}
\title{Default summary function for dpGLM class}
\usage{
\method{summary}{dpGLM}(object, ...)
}
\arguments{
\item{object}{a \code{dpGLM} object returned by the function \code{hdpGLM}}

\item{...}{The additional parameters accepted are:

           true.beta: (see \link{plot.dpGLM})}
}
\description{
This function provides a summary of the MCMC samples from the dpGLM model
}
\details{
Data points are assigned to clusters according to the highest estimated probability of belonging to that cluster
}
